E2E: move Me smoke spec alongside the other dashboard specs
